--- src/main/java/uk/gov/hmcts/reform/ccd/document/am/apihelper/SecurityHelper.java
@@ -1,29 +1,0 @@
-package uk.gov.hmcts.reform.ccd.document.am.apihelper;
-
-import feign.Feign;
-import feign.jackson.JacksonEncoder;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cloud.openfeign.support.SpringMvcContract;
-import org.springframework.context.annotation.Bean;
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi;
-import uk.gov.hmcts.reform.authorisation.generators.ServiceAuthTokenGenerator;
-
-@Component
-public class SecurityHelper {
-    @Bean
-    public ServiceAuthorisationApi generateServiceAuthorisationApi(@Value("${idam.s2s-auth.url}") final String s2sUrl) {
-        return Feign.builder()
-                    .encoder(new JacksonEncoder())
-                    .contract(new SpringMvcContract())
-                    .target(ServiceAuthorisationApi.class, s2sUrl);
-    }
-
-    @Bean
-    public ServiceAuthTokenGenerator authTokenGenerator(
-        @Value("${idam.s2s-auth.totp_secret}") final String secret,
-        @Value("${idam.s2s-auth.microservice}") final String microService,
-        final ServiceAuthorisationApi serviceAuthorisationApi) {
-        return new ServiceAuthTokenGenerator(secret, microService, serviceAuthorisationApi);
-    }
-}
