+++ src/main/java/uk/gov/hmcts/reform/ccd/document/am/configuration/SecurityConfiguration.java
@@ -1,6 +1,11 @@
 package uk.gov.hmcts.reform.ccd.document.am.configuration;
 
+import feign.Feign;
+import feign.jackson.JacksonEncoder;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cloud.openfeign.support.SpringMvcContract;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.ProviderManager;
@@ -12,6 +17,8 @@ import uk.gov.hmcts.reform.auth.checker.core.RequestAuthorizer;
 import uk.gov.hmcts.reform.auth.checker.core.service.Service;
 import uk.gov.hmcts.reform.auth.checker.core.user.User;
 import uk.gov.hmcts.reform.auth.checker.spring.serviceanduser.AuthCheckerServiceAndUserFilter;
+import uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi;
+import uk.gov.hmcts.reform.authorisation.generators.ServiceAuthTokenGenerator;
 
 import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
 
@@ -62,5 +69,21 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
             .anyRequest()
             .authenticated();
 
+    }
+
+    @Bean
+    public ServiceAuthorisationApi generateServiceAuthorisationApi(@Value("${idam.s2s-auth.url}") final String s2sUrl) {
+        return Feign.builder()
+                    .encoder(new JacksonEncoder())
+                    .contract(new SpringMvcContract())
+                    .target(ServiceAuthorisationApi.class, s2sUrl);
+    }
+
+    @Bean
+    public ServiceAuthTokenGenerator authTokenGenerator(
+        @Value("${idam.s2s-auth.totp_secret}") final String secret,
+        @Value("${idam.s2s-auth.microservice}") final String microService,
+        final ServiceAuthorisationApi serviceAuthorisationApi) {
+        return new ServiceAuthTokenGenerator(secret, microService, serviceAuthorisationApi);
     }
 }
